release 0.0.6-beta source code for go

+++ services/rds/v3/model/model_get_restore_time_response_restore_time.go
@@ -7,9 +7,9 @@ import (
 )
 
 type GetRestoreTimeResponseRestoreTime struct {
-	StartTime int32 `json:"start_time"`
+	StartTime int64 `json:"start_time"`
 
-	EndTime int32 `json:"end_time"`
+	EndTime int64 `json:"end_time"`
 }
 
 func (o GetRestoreTimeResponseRestoreTime) String() string {
